Changes moved to MonetizationTeams
